+++ scilab/modules/ast/includes/types/mlist.hxx
@@ -49,7 +49,7 @@ public :
         return false;
     }
 
-    bool                            getMemory(int* _piSize, int* _piSizePlusType);
+    bool                            getMemory(long long* _piSize, long long* _piSizePlusType);
 
     virtual bool invoke(typed_list & in, optional_list & opt, int _iRetCount, typed_list & out, const ast::Exp & e) override;
