+++ scilab/modules/ast/includes/types/tlist.hxx
@@ -76,7 +76,7 @@ public :
 
     bool                            toString(std::wostringstream& ostr);
 
-    bool                            getMemory(int* _piSize, int* _piSizePlusType);
+    bool                            getMemory(long long* _piSize, long long* _piSizePlusType);
 
 private :
 };
