+++ scilab/modules/ast/src/cpp/types/graphichandle.cpp
@@ -206,7 +206,7 @@ bool GraphicHandle::transpose(InternalType *& out)
     return type_traits::transpose(*this, out);
 }
 
-bool GraphicHandle::getMemory(int* _piSize, int* _piSizePlusType)
+bool GraphicHandle::getMemory(long long* _piSize, long long* _piSizePlusType)
 {
     *_piSize = sizeof(long long)*getSize();
     *_piSizePlusType = *_piSize + sizeof(GraphicHandle);
