+++ scilab/modules/ast/src/cpp/types/sparse.cpp
@@ -520,7 +520,7 @@ Sparse::Sparse(int rows, int cols, int nonzeros, int* inner, int* outer, double*
 }
 
 
-bool Sparse::getMemory(int *_piSize, int* _piSizePlusType)
+bool Sparse::getMemory(long long *_piSize, long long* _piSizePlusType)
 {
     *_piSize = nonZeros()*sizeof(double)*(isComplex() ? 2 : 1);
     *_piSizePlusType = *_piSize + sizeof(*this);
@@ -3227,7 +3227,7 @@ SparseBool::SparseBool(int rows, int cols, int trues, int* inner, int* outer)
     matrixBool->resizeNonZeros(trues);
 }
 
-bool SparseBool::getMemory(int *_piSize, int* _piSizePlusType)
+bool SparseBool::getMemory(long long *_piSize, long long* _piSizePlusType)
 {
     *_piSize = nbTrue()*sizeof(bool);
     *_piSizePlusType = *_piSize + sizeof(*this);
