+++ scilab/modules/ast/src/cpp/types/struct.cpp
@@ -106,14 +106,14 @@ Struct::Struct(Struct *_oStructCopyMe)
 #endif
 }
 
-bool Struct::getMemory(int* _piSize, int* _piSizePlusType)
+bool Struct::getMemory(long long* _piSize, long long* _piSizePlusType)
 {
     *_piSize = 0;
     *_piSizePlusType = 0;
     SingleStruct** p = get();
     for (int i = 0; i < getSize(); i++)
     {
-        int piS, piSPT;
+        long long piS, piSPT;
         if (p[i]->getMemory(&piS, &piSPT))
         {
             *_piSize += piS;
